requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
boto3==1.20.50
    # via -r requirements.in
botocore==1.23.50
    # via
    #   boto3
    #   s3transfer
certifi==2021.10.8
    # via
    #   pipenv
    #   requests
charset-normalizer==2.0.11
    # via requests
distlib==0.3.4
    # via virtualenv
django==4.0.2
    # via image
filelock==3.4.2
    # via virtualenv
idna==3.3
    # via requests
image==1.5.33
    # via -r requirements.in
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
numpy==1.22.2
    # via -r requirements.in
pillow==9.0.1
    # via image
pipenv==2022.1.8
    # via -r requirements.in
platformdirs==2.4.1
    # via virtualenv
python-dateutil==2.8.2
    # via botocore
requests==2.27.1
    # via -r requirements.in
s3transfer==0.5.1
    # via boto3
six==1.16.0
    # via
    #   image
    #   python-dateutil
    #   virtualenv
sqlparse==0.4.2
    # via django
urllib3==1.26.8
    # via
    #   botocore
    #   requests
virtualenv==20.13.1
    # via pipenv
virtualenv-clone==0.5.7
    # via pipenv

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
